new functions and assembler
